CSS overhaul
Bitchass css never doin what i tell it to do
