+++ index.html
@@ -4,15 +4,15 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="styles/style.css" rel="stylesheet">
-    <script src="./script.js" defer></script>
+    <script src="./script.js" type="module" defer></script>
     <title>Ticketing System</title>
     <link rel="logo" href="my-app/src/images/Raygun_Logo.png" type="image/x-icon">
 </head>
 <body>
     <main>
-        <div class="center">
-            <h1 style="margin-top: 150px;">Welcome</h1>
-            <h2 style="margin-top: 1px;">Please Login</h2>
+        <div id="welcome-tags" class="center">
+            <h1>Welcome.</h1>
+            <h2>Please signup/login:</h2>
         </div>
 
         <div id="signup-form-container" class="center">
@@ -26,17 +26,16 @@
                 <label for="password"><b>Password</b></label>
                 <input type="password" placeholder="Enter Password" name="password" required>
 
-                <label for="role"><b>Role</b></label>
-                <input type="text" placeholder="Enter Role" name="role" required>
-
                 <div class="landing-buttons">
                     <button type="submit">Sign Up</button>
                     <button type="button" id="login">Login</button>
                 </div>
             </form>
+
+            <p id="signup-error" class="center"re></p>
         </div>
 
-        <p id="signup-error"></p>
+        
     </main>
 </body>
