+++ styles/style.css
@@ -3,13 +3,24 @@
     src: url("fonts/Lora/Lora-Regular.ttf");
 }
 
+@keyframes fadeIn {
+    0% {
+        opacity: 0;
+    }
+
+    100% {
+        opacity: 1;
+    }
+}
+
 * {
     font-family: 'lora-regular';
 }
 
 body {
-    background-color: #353333;
-    color: whitesmoke
+    background: fixed;
+    background-image: linear-gradient(white 75%, #353333 105%);
+    /* color: whitesmoke; */
 }
 
 #footer {
@@ -18,6 +29,7 @@ body {
     bottom: 0;
     width: 100%;
     text-align: center;
+    margin-bottom: 10px;
 }
 
 .center {
@@ -29,30 +41,42 @@ body {
 h1 {
     font-family: 'lora-regular';
     font-size: 80px;
+    /* border: 2px solid white; */
 }
 
-h2{
+h2 {
     font-family: 'lora-regular';
-    font-size: 20px;
+    font-size: 25px;
+    /* border: 2px solid white; */
 }
 
 #signup-form-container {
-    width: 25%;
-    height: 500px;
-    border: 1px solid #DBF9B8;
-    border-radius: 2%;
-    margin-top: auto;
-    background-color: #CAE2BC;
-    box-shadow: 7px 7px 7px;
+    width: 40%;
+    float: right;
+    margin-right: 10%;
+    /* margin-top: auto; */
+    /* border: 2px solid white; */
+    margin-top: 15%;
+}
+
+#welcome-tags {
+    width: 40%;
+    float: left;
+    margin-left: 10%;
+    margin-top: 15%;
+    /* border: 2px solid white; */
 }
 
 .signup-form {
-    padding: 10px;
+    border-radius: 2%;
+    padding: 40px;
     width: 50%;
-    max-height: 50%;
+    max-height: 100%;
     margin: auto;
-    margin-top: 25%;
     text-align: center;
+    /* border: 1px solid #DBF9B8; */
+    background-color: #92AFD7;
+    box-shadow: 7px 7px 7px;
 
 }
 
@@ -61,10 +85,19 @@ form input {
     padding: 5px;
     border: 1px solid black;
     margin-bottom: 30px;
+    /* margin-right: 30%; */
 }
 
 form label {
     background-color: #ffffff;
-
+    padding: 5px;
     border-radius: 5px;
+    margin-right: 25px;
+}
+
+#signup-error {
+    animation: fadeIn ease 1s;
+    animation-iteration-count: 1;
+    animation-fill-mode: forwards;
+    padding-top: 20px;
 }
